fixes csharp1-h6

+++ CSharp-Fundamentals/06-Loops/10-OddAndEvenProduct/App.cs
@@ -7,10 +7,10 @@
     {
         private static void Main()
         {
-            var n = int.Parse(Console.ReadLine());
-            var numbers = Console.ReadLine().Split(' ').Select(int.Parse).ToArray();
-            var oddProductValue = 1;
-            var evenProductValue = 1;
+            var n = long.Parse(Console.ReadLine());
+            var numbers = Console.ReadLine().Split(' ').Select(long.Parse).ToArray();
+            var oddProductValue = 1L;
+            var evenProductValue = 1L;
             for (var i = 0; i < n; i++)
             {
                 var currentNumber = numbers[i];
